[Refactor]: split webhook gateway modules
